Added some extra support to the Color class, such as supporting HSV color space and making sure everything uses 32-bit floating point values.

+++ src/ColorTransform.hpp
@@ -40,19 +40,19 @@ namespace KRE
 		explicit ColorTransform(const variant& node);
 		explicit ColorTransform(const Color& color);
 		explicit ColorTransform(int ar, int ag, int ab, int aa);
-		explicit ColorTransform(double mr, double mg, double mb, double ma, double ar, double ag, double ab, double aa);
+		explicit ColorTransform(float mr, float mg, float mb, float ma, float ar, float ag, float ab, float aa);
 		explicit ColorTransform(int mr, int mg, int mb, int ma, int ar, int ag, int ab, int aa);
 		~ColorTransform();
 
-		double mulRed() const { return mul_rgba_[0]; }
-		double mulGreen() const { return mul_rgba_[1]; }
-		double mulBlue() const { return mul_rgba_[2]; }
-		double mulAlpha() const { return mul_rgba_[3]; }
+		float mulRed() const { return mul_rgba_[0]; }
+		float mulGreen() const { return mul_rgba_[1]; }
+		float mulBlue() const { return mul_rgba_[2]; }
+		float mulAlpha() const { return mul_rgba_[3]; }
 
-		double addRed() const { return add_rgba_[0]; }
-		double addGreen() const { return add_rgba_[1]; }
-		double addBlue() const { return add_rgba_[2]; }
-		double addAlpha() const { return add_rgba_[3]; }
+		float addRed() const { return add_rgba_[0]; }
+		float addGreen() const { return add_rgba_[1]; }
+		float addBlue() const { return add_rgba_[2]; }
+		float addAlpha() const { return add_rgba_[3]; }
 
 		void setAddRed(int ar);
 		void setAddGreen(int ag);
@@ -71,8 +71,8 @@ namespace KRE
 		std::string toString() const;
 	private:
 		DECLARE_CALLABLE(ColorTransform);
-		double mul_rgba_[4];
-		double add_rgba_[4];
+		float mul_rgba_[4];
+		float add_rgba_[4];
 	};
 
 	ColorTransform operator+(const ColorTransform& a, const ColorTransform& b);
